feat: E2E testing infrastructure — scribe-test crate and Docker containers
New scribe-test crate with a daemon+CLI architecture for end-to-end
testing Scribe in Docker containers. A persistent daemon holds the
IPC connection to scribe-server, buffering per-session state. CLI
subcommands talk to the daemon via a local Unix socket, enabling
bash test scripts to drive the terminal programmatically.

Crate modules:
- daemon: async event loop, message dispatch, 64KB output ring
  buffer, wait conditions (output regex, CWD, idle, exit) via
  tokio::Notify
- cmd_socket: DaemonRequest/DaemonResponse protocol over local socket
- ipc: async client to scribe-server using scribe-common framing
- server: start/stop scribe-server process lifecycle
- render: CPU PNG renderer using cosmic-text + png crate (no GPU)
- session/input/wait/assert/capture: CLI subcommand implementations

Protocol changes:
- Add RequestSnapshot to ClientMessage
- Implement CreateWorkspace handler in dispatch_message
- Make snapshot_term helpers pub for reuse from ipc_server
- Add FromStr and to_full_string to SessionId/WorkspaceId

Docker:
- Functional container (ubuntu:24.04 + server + test harness, ~120MB)
- Visual container (adds Xvfb + Mesa llvmpipe + client, ~400MB)
- Entrypoints with trap-based cleanup
- Smoke test validating server->session->I/O->screenshot flow

+++ crates/scribe-common/src/screen.rs
@@ -10,9 +10,13 @@ pub struct ScreenCell {
 }
 
 /// Terminal color representation.
+///
+/// `Named` uses `u16` because `alacritty_terminal::NamedColor` has variants
+/// above 255 (e.g. `Foreground = 256`, `Background = 257`, `Cursor = 258`,
+/// `DimBlack = 259`, …). A `u8` would silently truncate these to 0–15.
 #[derive(Debug, Clone, Copy, Serialize, Deserialize)]
 pub enum ScreenColor {
-    Named(u8),
+    Named(u16),
     Indexed(u8),
     Rgb { r: u8, g: u8, b: u8 },
 }
